+++ module/actor/actor.mjs
@@ -747,6 +747,27 @@ export class HeroSystem6eActor extends Actor {
         return result;
     }
 
+    // Abort effect - If in combat and not our turn then this must be an abort unless holding an action
+    /**
+     * Is there combat and is it the actor's turn to act?
+     *
+     * @returns {boolean}
+     */
+    needsToAbortToAct() {
+        const currentCombatActorId = game.combat?.combatants.find(
+            (combatant) => combatant.tokenId === game.combat.current?.tokenId,
+        )?.actorId;
+        const thisActorsCombatTurn =
+            game.combat?.active && currentCombatActorId != undefined && currentCombatActorId === this.id;
+        const thisActorHoldingAnAction = this.statuses.has("holding");
+
+        if (game.combat?.active && !thisActorsCombatTurn && !thisActorHoldingAnAction) {
+            return true;
+        }
+
+        return false;
+    }
+
     // When stunned, knockedout, etc you cannot act
     canAct(uiNotice, event) {
         // Bases can always act (used for token attacher)
@@ -773,6 +794,13 @@ export class HeroSystem6eActor extends Actor {
             result = false;
         }
 
+        // Is not actor's turn to act
+        if (this.needsToAbortToAct()) {
+            badStatus.push("NOT THE ACTIVE COMBATANT");
+            result = false;
+        }
+
+        // No speed?
         if (parseInt(this.system.characteristics.spd?.value || 0) < 1) {
             if (uiNotice) badStatus.push("SPD1");
             result = false;
@@ -2323,24 +2351,6 @@ export class HeroSystem6eActor extends Actor {
     }
 
     async addAttackPlaceholders() {
-        // Maneuver Strength Placeholder
-        // PH: FIXME: Figure out how to hide this (has name "__strengthPlaceholderWeapon") in the UI
-        const strengthPlaceholderItemContent = `<POWER XMLID="__STRENGTHDAMAGE" ID="1709333792635" BASECOST="0.0" LEVELS="1" ALIAS="__InternalStrengthPlaceholder" POSITION="4" MULTIPLIER="1.0" GRAPHIC="Burst" COLOR="255 255 255" SFX="Default" SHOW_ACTIVE_COST="Yes" INCLUDE_NOTES_IN_PRINTOUT="Yes" NAME="__InternalStrengthPlaceholder" INPUT="PD" USESTANDARDEFFECT="No" QUANTITY="1" AFFECTS_PRIMARY="No" AFFECTS_TOTAL="Yes"></POWER>`;
-        const strengthPlaceholderItemData = HeroSystem6eItem.itemDataFromXml(strengthPlaceholderItemContent, this);
-        const strengthPlaceholderItem = this.id
-            ? await HeroSystem6eItem.create(strengthPlaceholderItemData, {
-                  parent: this,
-              })
-            : new HeroSystem6eItem(strengthPlaceholderItemData, {
-                  parent: this,
-              });
-
-        // Work around if temporary actor
-        if (!this.id) {
-            this.items.set(strengthPlaceholderItem.name, strengthPlaceholderItem);
-        }
-        await strengthPlaceholderItem._postUpload();
-
         // Maneuver Weapon Placeholder
         // PH: FIXME: Figure out how to hide this (has name "__InternalManeuverPlaceholderWeapon") in the UI
         const maneuverWeaponPlaceholderItemContent = `<POWER XMLID="__STRENGTHDAMAGE" ID="1709333792633" BASECOST="0.0" LEVELS="1" ALIAS="__InternalManeuverPlaceholderWeapon" POSITION="4" MULTIPLIER="1.0" GRAPHIC="Burst" COLOR="255 255 255" SFX="Default" SHOW_ACTIVE_COST="Yes" INCLUDE_NOTES_IN_PRINTOUT="Yes" NAME="__InternalManeuverPlaceholderWeapon" INPUT="PD" USESTANDARDEFFECT="No" QUANTITY="1" AFFECTS_PRIMARY="No" AFFECTS_TOTAL="Yes"></POWER>`;