+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Releases
 
-## Version 4.0.25 20250329 [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)
+## Version 4.0.26 So Far... [Hero System 6e (Unofficial) v2](https://github.com/dmdorman/hero6e-foundryvtt)
+
+- Support Naked Advantages for attack powers.
+- Support advantages for Hand-to-Hand Attacks.
+
+## Version 4.0.25 20250329
 
 - Fixed maps failing to load when older maps used older "dead" status [#2099](https://github.com/dmdorman/hero6e-foundryvtt/issues/2099)
 - Initial support for SURFACE templates [#2055](https://github.com/dmdorman/hero6e-foundryvtt/issues/2055)